refactor: move GraphQLSchemaExport to debug directory
  - Relocate GraphQLSchemaExport component to components/debug/ for better organization
  - Update import path in App.tsx to reflect new location
